Relocated filter.data for use in upcoming tests.
